Add files via upload
01 
I have extended and implemented Massi's work on the right sidebar to production. 
Here I have extended the style 'new_font' (Massi, please use better names in the future) as below and applied to keywords and numbering respectively. 

Note instead of pixels for font, we would like to use 'em' which is equivalent to percentage (1em = 100%) and refers itself to the baseline (inherited) font size. This is the best for responsive layouts.

I struggled somewhat and then had to remove text-align:left as it and .main-menu{width:50px} because it was interfering with rest of the elements. I know the keywords are now too much to the left, but will take it up later, at least it is not interfering with other elements.

As Andrea suggested, I have made made the number bold.

.NavRightKeywordFont {
font-family: Ubuntu; 
font-size: 1.0em; 
} 

.NavRightNumberFont {
font-family: Ubuntu; 
font-size: 1.0em; 
font-weight:700; 
}